+++ src/briefy/ws/__init__.py
@@ -1,14 +1,22 @@
 """Briefy microservices helper."""
 from .initialization import initialize  # noqa
-from pyramid.renderers import JSONP
+from .renderer import JSONRenderer
+from briefy.ws.config import JWT_EXPIRATION
+from briefy.ws.config import JWT_SECRET
+from briefy.ws.auth import groupfinder
+from briefy.ws.auth import user_factory
+from pyramid.authorization import ACLAuthorizationPolicy
 
 import os
+import logging
 
 
 CORS_POLICY = {
     'origins': ('*', ),
     'headers': ('X-Locale', )
 }
+
+logger = logging.getLogger(__name__)
 
 
 def expandvars_dict(settings):
@@ -24,14 +32,25 @@ def expandvars_dict(settings):
 def includeme(config):
     """Configuration to be included by other services."""
 
-    # add default renderer
-    config.add_renderer('jsonp', JSONP(param_name='callback'))
-
     # Setup cornice.
     config.include("cornice")
 
+    # add default renderer
+    config.add_renderer('json', JSONRenderer())
+
     # Per-request transaction.
     config.include("pyramid_tm")
+
+    # config jwt
+    config.set_authorization_policy(ACLAuthorizationPolicy())
+    config.include('pyramid_jwt')
+    config.set_jwt_authentication_policy(private_key=JWT_SECRET,
+                                         expiration=int(JWT_EXPIRATION),
+                                         callback=groupfinder,
+                                         )
+
+    # add authenticated user map as request attribute
+    config.add_request_method(user_factory, 'user', reify=True)
 
     # Scan views.
     config.scan("briefy.ws.views")