+++ tests/testapp/views.py
@@ -1,8 +1,11 @@
 """Testapp views"""
 from briefy.ws import CORS_POLICY
+from briefy.ws.auth import validate_jwt_token
 from cornice import Service
+from pyramid.view import view_config
 
 import pyramid.httpexceptions as http_exc
+
 
 dummy = Service(
     name='Dummy service',
@@ -29,7 +32,44 @@ success = Service(
     description="Simplest app"
 )
 
+
 @success.get()
 def get_info(request):
     """Returns Hello in JSON."""
     return {'Hello': 'World'}
+
+
+protected = Service(
+    name='protected',
+    path='/protected',
+    description="Protected view"
+)
+
+
+@view_config(route_name='login', renderer='json')
+def login_view(request):
+    user = {
+        "locale": "en_GB",
+        "id": "669a99c2-9bb3-443f-8891-e600a15e3c10",
+        "fullname": "Person Lastname",
+        "first_name": "Person",
+        "email": "person@gmail.com",
+        "last_name": "Lastname",
+        "groups": ["g:briefy_pms", "g:briefy_qa"]
+    }
+    token = request.create_jwt_token(user.get('id'), **user)
+    result = dict(token=token,
+                  status='200',
+                  message='Authentication success',
+                  provider='email',
+                  user=user)
+    return result
+
+
+@protected.get(validators=validate_jwt_token)
+def get_protected(request):
+    """Returns protected information when the user is authenticated."""
+    return {'status': 'success',
+            'message': 'Protected information.',
+            'user': request.user.to_dict(),
+            }